+++ src/main/java/warmup/two/WarmupTwoFirstSet.java
@@ -45,13 +45,21 @@ public class WarmupTwoFirstSet {
     }
 
     /**
-     * Convert the number of "xx" in the given string. We'll say that overlapping is allowed, so
-     * "xxx" contains 2 "xx"
+     * Count the number of "xx" in the given string. We'll say that overlapping is allowed, so
+     * "xxx" contains 2 "xx".
      *
      * @param str
      * @return
      */
     public int countXX(String str) {
-        return 1;
+        int count = 0;
+
+        for (int i = 0; i < str.length()-1; i++) {
+            if (str.substring(i, i+2).equals("xx")) {
+                count++;
+            }
+        }
+
+        return count;
     }
 }